Distro related files reorga
