+++ react/src/views/Distributions/DistroSpotsView/DistroSpotsView.tsx
