+++ react/src/views/Distributions/DistroSpotsView/components/DistroSpots.stories.tsx
