+++ react/src/views/Distributions/DistroSpotsView/components/DistroSpots.tsx
